Make evaluation_kernels only include FEEvaluationData file
